Moves content from a few files to warrior.js
